whisper-wrapper/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai-whisper==20230918
numpy==1.24.3
torch==2.0.1